Rename utils.hpp to avoid name conflict
