+++ src/frontends/tests/frontend/shared/src/json_config.cpp
@@ -11,7 +11,7 @@
 #include <ostream>
 
 #include "common_test_utils/file_utils.hpp"
-#include "utils.hpp"
+#include "frontend_test_utils.hpp"
 
 using namespace ov::frontend;
