+++ src/frontends/tests/frontend/shared/src/partial_shape.cpp
@@ -4,7 +4,7 @@
 
 #include "partial_shape.hpp"
 
-#include "utils.hpp"
+#include "frontend_test_utils.hpp"
 
 using namespace ngraph;
 using namespace ov::frontend;
