+++ src/frontends/tests/frontend/shared/src/set_element_type.cpp
@@ -4,7 +4,7 @@
 
 #include "set_element_type.hpp"
 
-#include "utils.hpp"
+#include "frontend_test_utils.hpp"
 
 using namespace ngraph;
 using namespace ov::frontend;
