+++ src/frontends/tests/frontend/shared/src/telemetry.cpp
@@ -4,8 +4,8 @@
 
 #include "telemetry.hpp"
 
+#include "frontend_test_utils.hpp"
 #include "openvino/frontend/extension/telemetry.hpp"
-#include "utils.hpp"
 
 using namespace ov::frontend;
